refactor(maintenance): extract maintenance shell component
